Split factories into multiple package
 - django 1.11 is less tolerant with importing models that belong to
   no installed apps, hence it fails if the maps is not installed
