+++ tests/maps/conftest.py
@@ -1,6 +1,6 @@
 from pytest_factoryboy import register
 
-from adhocracy4.test import factories
+from adhocracy4.test.factories import maps as map_factories
 
 
-register(factories.AreaSettingsFactory)
+register(map_factories.AreaSettingsFactory)
